Move error to its own class
